requirements.txt
ftfy==6.1.3
loralib==0.1.1
numpy==1.22.3
Pillow==10.0.0
regex==2022.10.31
scikit_learn==1.3.0
scipy==1.9.1
tabulate==0.9.0
timm==0.9.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.63.0
transformers==4.28.1
wandb==0.13.4
yacs==0.1.8
